app/requirements.txt
dash
dash-core-components
dash-html-components
dash-renderer
dash-table
Flask-SQLAlchemy
python-dotenv
dash_google_auth
psycopg2
pandas
sty
psycopg2-binary
